+++ Cargo.toml
@@ -9,7 +9,7 @@ edition = "2021"
 rfd = "0.11.3"
 tracing = "0.1.29"
 tracing-subscriber = "0.3.3"
-eframe = "0.16"
+eframe = "0.21.3"
 urlencoding = "2.1.0"
 rmp = "0.8.10"
 webbrowser = "0.5.5"
